fix bake bug

+++ shader_utils.py
@@ -188,8 +188,8 @@ def build_model(objects, get_uv = False):
 
           
         uvs = np.asarray(uvs)
-        uv_indices = np.asarray(uv_indices)
-        loop_indices = np.asarray(loop_indices) 
+        uv_indices = np.asarray(uv_indices, dtype = np.int32)
+        loop_indices = np.asarray(loop_indices, dtype = np.int32) 
 
     camera_loc = camera.location
